# Improve performance of TwoBitReader