#### Summary
Quick fix for #150 : Random read of 2bit is quite slower than FASTA
Now it got about **20x ~ 35x** faster than master

#### Benchmark

https://github.com/chrovis/cljam/blob/7921def33b712c099fd6cac94fe1e98ca2a94c76/bench/cljam/io/sequence_bench.clj#L7-L9

https://github.com/chrovis/cljam/blob/7921def33b712c099fd6cac94fe1e98ca2a94c76/bench/cljam/io/sequence_bench.clj#L15-L22

https://github.com/chrovis/cljam/blob/7921def33b712c099fd6cac94fe1e98ca2a94c76/bench/cljam/io/sequence_bench.clj#L28-L37

##### master
```clj
read-all-sequences-test (sequence_bench.clj:7)

  ["test-resources/fasta/medium.fa" {}]
  time: 78.420780 ms, sd: 7.932527 µs

  ["test-resources/fasta/medium.fa" {:mask? true}]
  time: 80.570329 ms, sd: 27.922591 µs

  ["test-resources/twobit/medium.2bit" {}]
  time: 121.734670 ms, sd: 1.247978 µs

  ["test-resources/twobit/medium.2bit" {:mask? true}]
  time: 125.450115 ms, sd: 19.756194 µs

read-sequence-once-test (sequence_bench.clj:15)

  ["test-resources/fasta/medium.fa" {}]
  time: 651.141680 µs, sd: 0.492259 ns

  ["test-resources/fasta/medium.fa" {:mask? true}]
  time: 688.146704 µs, sd: 1.269185 ns

  ["test-resources/twobit/medium.2bit" {}]
  time: 9.856713 ms, sd: 26.594884 ns

  ["test-resources/twobit/medium.2bit" {:mask? true}]
  time: 10.111649 ms, sd: 130.142010 ns

read-sequence-test (sequence_bench.clj:28)

  ["test-resources/fasta/medium.fa" {}]
  time: 236.486612 µs, sd: 0.017035 ns

  ["test-resources/fasta/medium.fa" {:mask? true}]
  time: 248.022497 µs, sd: 0.039226 ns

  ["test-resources/twobit/medium.2bit" {}]
  time: 9.423483 ms, sd: 11.572672 ns

  ["test-resources/twobit/medium.2bit" {:mask? true}]
  time: 9.261297 ms, sd: 28.707967 ns
```

##### PR
```clj
read-all-sequences-test (sequence_bench.clj:7)

  ["test-resources/fasta/medium.fa" {}]
  time: 79.265507 ms, sd: 1.977376 µs

  ["test-resources/fasta/medium.fa" {:mask? true}]
  time: 80.113221 ms, sd: 2.527757 µs

  ["test-resources/twobit/medium.2bit" {}]
  time: 3.511812 ms, sd: 1.397351 ns

  ["test-resources/twobit/medium.2bit" {:mask? true}]
  time: 5.465990 ms, sd: 9.133830 ns

read-sequence-once-test (sequence_bench.clj:15)

  ["test-resources/fasta/medium.fa" {}]
  time: 652.469276 µs, sd: 0.422977 ns

  ["test-resources/fasta/medium.fa" {:mask? true}]
  time: 653.855513 µs, sd: 0.440200 ns

  ["test-resources/twobit/medium.2bit" {}]
  time: 376.351904 µs, sd: 0.103506 ns

  ["test-resources/twobit/medium.2bit" {:mask? true}]
  time: 514.818163 µs, sd: 0.014057 ns

read-sequence-test (sequence_bench.clj:28)

  ["test-resources/fasta/medium.fa" {}]
  time: 240.530200 µs, sd: 0.021139 ns

  ["test-resources/fasta/medium.fa" {:mask? true}]
  time: 240.975009 µs, sd: 0.015479 ns

  ["test-resources/twobit/medium.2bit" {}]
  time: 263.236439 µs, sd: 0.010704 ns

  ["test-resources/twobit/medium.2bit" {:mask? true}]
  time: 408.126779 µs, sd: 0.025648 ns
```

#### Changes
- Add bench scripts
- Rewrite `cljam.io.twobit.reader.TwoBitReader` leveraging `java.nio.MappedByteBuffer`
- Add `read-seq-summaries` fn for quick access to `:name` and `:len` without loading indices

#### Tests

- `lein check` 🆗
- `lein test :all` 🆗
- `lein eastwood` 🆗