@alumi 2018-10-31
I don't have detailed profiles for each change, but I guess that the root cause was using a `java.io.RandomAccessFile` without buffering. Replacing it with a `java.io.BufferedInputStream` substantially improved the performance. We can use a `java.nio.MappedByteBuffer` here to make it even better because size of 2bit v0 file never exceed 2GB. (v1 does, but we're not supporting it)
Additionally, some minor changes slightly speeded up constructing result String
- Decode 4 characters at a time and slice buffer later
- Use [[C instead of a vector of Strings for decoding
- Use `java.util.TreeMap` to avoid linear searching of ambiguous regions and masked regions
- Use `java.util.HashMap` to find sequence headers with chromosome names
- Use unchecked maths