@athos 2018-11-01
Thanks for the details. I see, that's interesting.

> I guess that the root cause was using a java.io.RandomAccessFile without buffering. Replacing it with a java.io.BufferedInputStream substantially improved the performance. We can use a java.nio.MappedByteBuffer here to make it even better because size of 2bit v0 file never exceed 2GB.

I misunderstood `RandomAccessFile` does always memory-map the file content, but actually it doesn't unless the `map` method is called via its channel. I didn't get why `FASTAReader` is so much faster in spite of the use of `RandomAccessFile` as `TwoBitReader` did. Now I see your point.

And I also noticed that `ByteBuffer` supports both byte orders natively, which may have contributed to not so small gain in performance as well.