@totakke 2018-11-02
I've enabled `Require pull request reviews before merging` to avoid such a mistake.

@alumi This PR was closed by my fault. Would you recreate same PR?